Actually split the graphics functionality to its own file.
There's still some rough edges, to be sanded off in the next commits, but this works as a complete split already.
